+++ home/CHANGELOG.md
@@ -1,6 +1,11 @@
-# Changelog
+# Changelog for `home` app
 
 ## [Unreleased]
+
+### Fixed
+
+- Recent live pages not showing up in `AppLandingPage`. 
+[#17](https://git.bebif.be/antabif/biodiversityaq/issues/17) ([ymgan](https://git.bebif.be/ymgan))
 
 ### Added